% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus.k.R
\name{clus.k}
\alias{clus.k}
\title{Calculate number of clusters which cluster::pam needs}
\usage{
clus.k(df)
}
\arguments{
\item{df}{a data frame.}
}
\description{
Calculate number of clusters which cluster::pam needs
}
\examples{
# example 1
df <- iris[,-5]; clus.k(df)

# example 2
data(wine, package = 'rattle'); df <- wine[,-1]; clus.k(df)

}
\author{
Chand Zhang
}
