% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clus.ward.R
\name{clus.ward}
\alias{clus.ward}
\title{Calculate Ward cluster tree}
\usage{
clus.ward(df, k, writetree = F)
}
\arguments{
\item{df}{a data frame.}

\item{k}{number of clusters.}

\item{writetree}{if TRUE, output the tree to working directory.}
}
\description{
Calculate Ward cluster tree
}
\examples{
# example 1
df <- iris[,-5]; clus.ward(df, 3)
table(df.gauss, iris$Species)

# example 2
data(wine, package = 'rattle'); df <- wine
clus.ward(df[,-1], k = 3)
table(df.ward, df$Type)

}
\author{
Chand Zhang
}
