% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr}
\alias{corr}
\title{Compute correlation matrix}
\usage{
corr(v, pcc = F)
}
\arguments{
\item{v}{a data.frame (or list) should be taken.}

\item{pcc}{if TRUE, ignore result of normality test and use Pearson correlation coefficient}
}
\description{
Compute Spearman and Pearson correlation matrix.
}
\examples{
# example 1
corr(mtcars[, 1:5])

# example 2
corr(mtcars[, 1:5], pcc = T)

}
