% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ward.R
\name{ward}
\alias{ward}
\title{Compute Ward cluster tree}
\usage{
ward(v, comm = F, k = NA, cex = 0.7, os = 0)
}
\arguments{
\item{v}{a numeric matrix, data frame.}

\item{comm}{if TRUE, data should belong to Community Ecology.}

\item{k}{number of clusters.}

\item{cex}{a numerical vector giving the amount by which plotting characters and symbols should be scaled relative to the default.}

\item{os}{distance between text and graphics.}
}
\description{
Compute Ward cluster tree
}
\examples{

set.seed(9); df <- iris[sample(150,45), ]; df <- df[order(df$Species), 1:4]
ward(v = df, comm = T, k = 3)

data(wine, package = 'rattle'); set.seed(6); df <- wine[sample(178,45), ]; df <- df[, -1]
ward(v = df, k = 3)

data(doubs, package = 'ade4'); spe <- doubs$fish[-8, ]
ward(spe, comm = T, k = 4)
}
