% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms.R
\name{kms}
\alias{kms}
\title{Compute k-means clustering}
\usage{
kms(v, comm = F, k = NA, check_overlap = F)
}
\arguments{
\item{v}{a numeric matrix, data frame.}

\item{comm}{if TRUE, data should be Community Ecology data(especially species data such as example 1).}

\item{k}{number of clusters.}

\item{check_overlap}{if TRUE, labels won't be overlap.}
}
\description{
Compute k-means clustering
}
\examples{
# example 1
data(doubs, package = 'ade4'); spe <- doubs$fish[-8, ]
kms(spe, comm = T); kms(spe, comm = T, k = 4)

# example 2(1:50; 51:100; 101:150)
set.seed(9); df <- iris[sample(150,70), ]; df <- df[order(df$Species), 1:4]
kms(df, comm = T); kms(df, comm = T, k = 3)

# example 3(1:59; 60:130; 131:178)
data(wine, package = 'rattle'); set.seed(6); df <- wine[sample(178,70), ]; df <- df[, -1]
kms(df, comm = T); kms(df, comm = T, k = 3, check_overlap = T)

}
