% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ward.R
\name{ward}
\alias{ward}
\title{Compute Ward cluster tree}
\usage{
ward(v, comm = F, k = NA, cex = 0.7, os = 0)
}
\arguments{
\item{v}{a numeric matrix, data frame.}

\item{comm}{if TRUE, data should be Community Ecology data(especially species data such as example 1).}

\item{k}{number of clusters.}

\item{cex}{a numerical vector giving the amount by which plotting characters and symbols should be scaled relative to the default.}

\item{os}{distance between text and graphics.}
}
\description{
Make sure that there is no newline after ward().
}
\examples{
# suggested method to read data
v <- readxl::read_xlsx('data.xlsx')
v1 <- as.matrix(v[, -1]); rownames(v1) <- v$colname
ward(v1)

# example 1
data(doubs, package = 'ade4'); spe <- doubs$fish[-8, ]
ward(spe, comm = T); ward(spe, comm = T, k = 4)

# example 2(1:50; 51:100; 101:150)
set.seed(9); df <- iris[sample(150,45), ]; df <- df[order(df$Species), 1:4]
ward(df, comm = T); ward(df, comm = T, k = 3)

# example 3(1:59; 60:130; 131:178)
data(wine, package = 'rattle'); set.seed(6); df <- wine[sample(178,45), ]; df <- df[, -1]
ward(df); ward(df, k = 3)
# meaning of standardization based on example 3
df$Alcohol <- 50*df$Alcohol + 100; ward(df, k = 3)

}
